feat(fetch-user-from-api): add users controller and user hooks
