+++ packages/react-ui/src/hooks/platform-user-hooks.ts
@@ -18,22 +18,4 @@ export const platformUserHooks = {
       },
     });
   },
-  useCurrentUser: () => {
-    const userId = authenticationSession.getCurrentUserId();
-    return useSuspenseQuery<UserWithMetaInformationAndProject | null, Error>({
-      queryKey: ['currentUser', userId],
-      queryFn: async () => {
-        if (!userId) {
-          return null;
-        }
-        const result = await platformUserApi.getCurrentUser();
-        return result;
-      },
-      staleTime: Infinity,
-    });
-  },
-  invalidateCurrentUser: (queryClient: QueryClient) => {
-    const userId = authenticationSession.getCurrentUserId();
-    queryClient.invalidateQueries({ queryKey: ['currentUser', userId] });
-  },
 };
