+++ packages/server/api/src/app/user/platform/platform-user-controller.ts
@@ -6,42 +6,15 @@ import {
     SeekPage,
     UpdateUserRequestBody,
     UserWithMetaInformation,
-    UserWithMetaInformationAndProject,
 } from '@activepieces/shared'
 import {
     FastifyPluginAsyncTypebox,
     Type,
 } from '@fastify/type-provider-typebox'
 import { StatusCodes } from 'http-status-codes'
-import { userIdentityService } from '../../authentication/user-identity/user-identity-service'
 import { userService } from '../user-service'
 
 export const platformUserController: FastifyPluginAsyncTypebox = async (app) => {
-
-    app.get('/me', GetCurrentUserRequest, async (req): Promise<UserWithMetaInformationAndProject> => {
-        const userId = req.principal.id
-        assertNotNullOrUndefined(userId, 'userId')
-
-        const user = await userService.getOneOrFail({ id: userId })
-        const identity = await userIdentityService(app.log).getOneOrFail({ id: user.identityId })
-
-        return {
-            id: user.id,
-            platformRole: user.platformRole,
-            status: user.status,
-            externalId: user.externalId,
-            created: user.created,
-            updated: user.updated,
-            platformId: user.platformId,
-            firstName: identity.firstName,
-            lastName: identity.lastName,
-            email: identity.email,
-            trackEvents: identity.trackEvents,
-            newsLetter: identity.newsLetter,
-            verified: identity.verified,
-            projectId: req.principal.projectId,
-        }
-    })
 
     app.get('/', ListUsersRequest, async (req) => {
         const platformId = req.principal.platform.id
@@ -76,18 +49,6 @@ export const platformUserController: FastifyPluginAsyncTypebox = async (app) => 
 
         return res.status(StatusCodes.NO_CONTENT).send()
     })
-}
-
-const GetCurrentUserRequest = {
-    schema: {
-        response: {
-            [StatusCodes.OK]: UserWithMetaInformationAndProject,
-        },
-    },
-    config: {
-        allowedPrincipals: [PrincipalType.USER],
-        scope: EndpointScope.PLATFORM,
-    },
 }
 
 const ListUsersRequest = {
